src/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.33.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic-settings==2.8.1
aiofiles==24.1.0
